+++ CHANGELOG.md
@@ -26,6 +26,7 @@ https://github.com/nwnxee/unified/compare/build8193.37.13...HEAD
 - Util: SetStartingLocation()
 - Object: GetLocalizedDescription(), SetLocalizedDescription()
 - Util: RawPrint()
+- Object: GetActionQueue(), GetActionQueuePosition()
 
 ### Changed
 - Damage: Added bRangedAttack to the NWNX_Damage_AttackEventData struct.